requirements/base.txt
# These are the reqs to build docs and run tests.
--no-binary :all:  # workaround for https://bitbucket.org/ned/coveragepy/issue/382/pip-install-coverage-uses-slower-pytracer
sphinx
django-nose
jinja2
redis
flake8
coverage
psycopg2
